chore: move tests related to twilio to a seperate subdirectory
