+++ src/test/endpoint/twilio/hangup.test.ts
@@ -4,7 +4,7 @@ import { Hangup } from '@/endpoint/hangup';
 import { fromHono } from 'chanfana';
 import { getConfig } from '@/lib/config';
 import { TwilioProvider } from '@/lib/providers/twilio';
-import { setupMockConfig } from '../utils/test-config';
+import { setupMockConfig } from '../../utils/test-config';
 
 // Mock only the config module to provide test configuration
 vi.mock('@/lib/config', () => ({
