+++ src/test/endpoint/twilio/incoming.test.ts
@@ -4,7 +4,7 @@ import { Incoming } from '@/endpoint/incoming';
 import { fromHono } from 'chanfana';
 import { getConfig } from '@/lib/config';
 import { TwilioProvider } from '@/lib/providers/twilio';
-import { createTestConfig, setupMockConfig } from '../utils/test-config';
+import { createTestConfig, setupMockConfig } from '../../utils/test-config';
 
 // Mock only the config module to provide test configuration
 vi.mock('@/lib/config', () => ({
